+++ rust/Cargo.toml
@@ -8,6 +8,7 @@ repository = "https://github.com/opendp/opendp"
 version = "0.0.0+development"
 authors = ["The OpenDP Project <info@opendp.org>"]
 edition = "2018"
+build = "build/main.rs"
 rust-version = "1.59.0" # MSRV via `cargo intall msrv && cargo msrv -- cargo check --features=untrusted,bindings-python`
 
 [dependencies]
@@ -26,9 +27,9 @@ lazy_static = { version = "1.4.0", optional = true }
 vega_lite_4 = { version = "0.6.0", optional = true }
 
 [build-dependencies]
-opendp_tooling = {path = "opendp_tooling"}
-syn = { version = "1.0" }
-proc-macro2 = { version = "1.0" }
+opendp_tooling = { path = "opendp_tooling", optional = true }
+syn = { version = "1.0", optional = true }
+proc-macro2 = { version = "1.0", optional = true }
 
 [features]
 default = ["use-openssl", "use-mpfr"]
@@ -49,6 +50,7 @@ test-plot = ["vega_lite_4"]
 
 ffi = ["lazy_static"]
 bindings-python = ["ffi"]
+derive = ["opendp_tooling", "opendp_derive/full", "syn", "proc-macro2"]
 
 [dev-dependencies]
 criterion = "0.4"
